moving merge methods to geotrellis.spark.mosaic
