+++ spark/src/main/scala/geotrellis/spark/ingest/package.scala
@@ -1,7 +1,6 @@
 package geotrellis.spark
 
 import geotrellis.proj4.CRS
-import geotrellis.raster.mosaic.MergeView
 import geotrellis.raster.resample.{NearestNeighbor, ResampleMethod}
 import geotrellis.spark.tiling.{LayoutDefinition, LayoutScheme}
 import geotrellis.vector._
@@ -72,16 +71,4 @@ package object ingest {
       tiler(tiles, rasterMetaData, resampleMethod)
     }
   }
-
-  implicit class withRddMergeMethods[K, TileType: MergeView](rdd: RDD[(K, TileType)])
-    extends RddMergeMethods[K, TileType](rdd)
-
-  implicit class withRddLayoutMergeMethods[K, TileType: MergeView](rdd: (RDD[(K, TileType)], LayoutDefinition))
-    extends RddLayoutMergeMethods[K, TileType](rdd)
-
-  implicit class withRasterRddMergeMethods[K, TileType: MergeView](rdd: RasterRDD[K])
-    extends RasterRddMergeMethods[K](rdd)
-
-  implicit class withMultiBandRasterRddMergeMethods[K, TileType: MergeView](rdd: MultiBandRasterRDD[K])
-    extends MultiBandRasterRddMergeMethods[K](rdd)
 }
