+++ spark/src/main/scala/geotrellis/spark/mosaic/MergeMethods.scala
@@ -1,0 +1,5 @@
+package geotrellis.spark.mosaic
+
+trait MergeMethods[T] {
+  def merge(other: T): T
+}
